requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
autoflake==1.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.4
Django==3.1.6
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
isort==5.7.0
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
promise==2.3
py==1.10.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
regex==2020.11.13
requests==2.25.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.0
